% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exoplanets-package.R
\docType{package}
\name{exoplanets-package}
\alias{exoplanets}
\alias{exoplanets-package}
\title{exoplanets: Exoplanet Datasets for Teaching Introductory Statistics}
\description{
Tidied datasets of confirmed exoplanets and their host stars, retrieved
from NASA's Exoplanet Archive via the \pkg{REXoplanets} package.
}
\section{Datasets}{

\itemize{
  \item \code{\link{planets}} - one row per confirmed exoplanet
        (\code{pscomppars} table)
  \item \code{\link{stars}} - reference-level stellar host data
        (\code{stellarhosts} table)
}
}

\section{Data dictionary}{

Machine-readable column dictionaries are shipped in \code{inst/extdata}:
\preformatted{
system.file("extdata", "data_dictionary_planets.csv", package = "exoplanets")
system.file("extdata", "data_dictionary_stars.csv",   package = "exoplanets")
system.file("extdata", "data_dictionary.md",          package = "exoplanets")
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/moderndive/exoplanets}
  \item Report bugs at \url{https://github.com/moderndive/exoplanets/issues}
}

}
\author{
\strong{Maintainer}: Chester Ismay \email{chester.ismay@gmail.com}

}
\keyword{internal}
